+++ packages/create-nuxe/lib/template/pages/index.vue
@@ -1,6 +1,6 @@
 <template>
   <div>
-    <h1>Welcome to your vuxe app</h1>
+    <h1>Welcome to your nuxe app</h1>
     <p>Current time: {{ now.toLocaleString() }}</p>
     <Hello name="World" />
   </div>