+++ packages/nuxe/lib/virtual-modules.d.ts
@@ -1,4 +1,4 @@
-declare module 'virtual:vuxe/layouts' {
+declare module 'virtual:nuxe/layouts' {
   const layouts: Record<string, import('vue').Component>
   export default layouts
 }